backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
SQLAlchemy==2.0.44
PyMySQL==1.1.2
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.2.1
pydantic==2.12.5
email-validator==2.3.0
python-multipart==0.0.20
gunicorn==23.0.0
pinecone==7.3.0
openai==1.78.1
